% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_exchange}
\alias{check_exchange}
\title{Helper functions}
\usage{
check_exchange(exchange = c("nse", "bse"), include_both = TRUE)
}
\arguments{
\item{exchange}{The value to be validated. Defaults to c("nse", "bse")}

\item{include_both}{Paramter to control whether to include "both" as an option for exchange or not. Defaults to TRUE}
}
\value{
The validated exchange value or en error if its not valid.
}
\description{
A function to validate the specified exchange value and get the exchange value if nothing is specified.
}
\examples{
IndianStocksR::: check_exchange()
}
