% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{compile_exchange_data}
\alias{compile_exchange_data}
\title{Compile the exchange files inside a directory}
\usage{
compile_exchange_data(data_path = "./data", exchange = c("both", "nse",
  "bse"), delete_component_files = TRUE, quiet = FALSE)
}
\arguments{
\item{data_path}{The path that contains the data files you want to compile. Deafults to the data folder in the current working directory.}

\item{exchange}{Specifies the exchange files you want to compile. One of "both", "nse" or "bse".
If both is specified, the nse component files are compiled into nse_compiled_latest_date.csv and
bse files are compiled into  bse_compiled_latest_date.csv.
Both the nse and bse compiled files are parsed and combined into both_compiled_latest_date.
Dedaults to "both".}

\item{delete_component_files}{Deletes all the individual files and retains only the compiled files.}
}
\value{
Returns the compiled dataframe and also has the returned dataframe written out as a csv in the folder it compiles.
}
\description{
This is used to compile all the individual files for different dates into compiled files.
}
\examples{
compile_exchange_data(data_path = "./data", exchange = "both")

}
