% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_validation}
\alias{date_validation}
\title{Helper functions}
\usage{
date_validation(date)
}
\arguments{
\item{date}{Any string that you want to see if it can be parsed as a date.}
}
\value{
The parsed date.
}
\description{
A date validator
}
\examples{

# date_validation("j765")

date_validation("2018-07-24")
date_validation(lubridate::today())
}
