% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_error_message}
\alias{download_error_message}
\title{Helper functions}
\usage{
download_error_message(date, exchange)
}
\arguments{
\item{date}{The date you want the data for.
The input should either be a date object or a something that lubridate::as_date can parse to be a
date. If a string is in the format yyyy-mm-dd, this generally works.}

\item{exchange}{Usually "nse" or "bse". But any string can be specified.}
}
\value{
The download error message
}
\description{
An error message constructor.
}
\examples{
download_error_message(lubridate::today(), "nse")
}
