% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_stocks}
\alias{download_stocks}
\title{Download stock data}
\usage{
download_stocks(date = lubridate::today(), exchange = c("nse", "bse"),
  dest_path = "./data", quiet = FALSE)
}
\arguments{
\item{date}{The date you want the data for.
The input should either be a date object or a something that lubridate::as_date can parse to be a
date. If a string is in the format yyyy-mm-dd, this generally works. Defaults to lubridate::today()}

\item{exchange}{Either "nse" or "bse". Defaults to "nse".}

\item{dest_path}{The location where you want to download the files to.
Defaults to the data folder in the current working directory.
If there is no data folder, one will be created.}

\item{quiet}{If TRUE, there will be no messages on the download status.}
}
\value{
he function does not return anything.
It just gets the files downloaded if data is available.
}
\description{
Downloads stock data from the specified exchanges on the specified date.
}
\examples{
download_stocks()

download_stocks(date = lubridate::today() - 3, exchange = "bse")
}
