% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_stocks_both}
\alias{download_stocks_both}
\title{Download stock data}
\usage{
download_stocks_both(date = lubridate::today(), ...)
}
\arguments{
\item{date}{The date you want the data for.
The input should either be a date object or a something that lubridate::as_date can parse to be a
date. If a string is in the format yyyy-mm-dd, this generally works. Defaults to lubridate::today()}

\item{...}{Arguments you would pass to the download stocks function}
}
\value{
The function does not return anything.
It just gets the files downloaded if data is available.
}
\description{
Used to download stocks from both exchanges on the date specified.
}
\examples{
download_stocks_both()

download_stocks_both(date = "2018-07-16", dest_path = ".")
}
