% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_date}
\alias{extract_date}
\title{Helper functions}
\usage{
extract_date(x)
}
\arguments{
\item{x}{A vector of filenames}
}
\value{
A vector of extracted dates from the filenames
}
\description{
Helper functions
}
\examples{

IndianStocksR:::extract_date(dir("./data"))
}
