% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safely_remove}
\alias{safely_remove}
\title{Helper functions}
\usage{
safely_remove(file)
}
\arguments{
\item{file}{The file with the location}
}
\value{
Deletes the file and returns nothing
}
\description{
Remove a file safely
}
\examples{
safely_remove("./data/file.csv")
}
